Update asm_processor, sorted boot_O1 into other folders, completed the fbdemo bootstrap, cleaned up undefined_syms

+++ src/buffers/heaps.c
@@ -2,4 +2,4 @@
 
 u8 gAudioHeap[0x138000];
 
-u32 gSystemHeap[UNK_SIZE];
+u8 gSystemHeap[UNK_SIZE];
